Move player's game replays into a separate component | Create PlayerOngoingTournaments component | Remove unused code
