+++ packages/client/src/components/PlayerOngoingTournaments.tsx
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+
+class PlayerOngoingTournaments extends Component {
+  render() {
+    return(
+      <h1>Player Ongoing Tournaments</h1>
+    )
+  }
+}
+
+export default PlayerOngoingTournaments;
